[REF] sale_rental_extension: remove rental_base-dependent code
Code that referenced rental_base (views, menus, actions, SaleOrder model,
StockPicking model, config settings, signature widget, delivery report
and tests) has been moved to rental_base_extension where it belongs.

This module now only extends sale_rental with:
- SaleOrderLine: rental_qty defaults, inline editing, qty sync
- sale.rental: kanban, calendar views and search enhancements
