+++ sale_rental_extension/README.rst
@@ -30,9 +30,8 @@ This module extends ``sale_rental`` to improve the rental positions experience:
 
 * Kanban and calendar views for the Rental Positions screen
 * Searchpanel with state filter and "Today" filter
-* Configurable signature terms for rental delivery orders
-  (Settings → Rental → Signature)
-* Custom signature dialog with "Confirm & Sign" button on rental pickings
+* Rental quantity defaults to 1 when adding a rental product
+* Rental quantity can be edited on confirmed orders (before pickup)
 * Catalan and Spanish translations
 
 .. IMPORTANT::
