+++ sale_rental_extension/__manifest__.py
@@ -3,9 +3,9 @@
 
 {
     "name": "Sale Rental Extension",
-    "summary": "Dedicated rental views, rental status tracking,"
-    " kanban, calendar, configurable signature terms and translations.",
-    "version": "16.0.1.4.0",
+    "summary": "Rental position kanban, calendar views,"
+    " rental quantity defaults and inline editing improvements.",
+    "version": "16.0.1.5.0",
     "development_status": "Alpha",
     "category": "Rental",
     "author": "NuoBiT Solutions SL",
@@ -15,17 +15,6 @@
         "sale_rental",
     ],
     "data": [
-        "report/report_deliveryslip.xml",
-        "views/sale_order_rental_views.xml",
-        "views/res_config_settings_views.xml",
         "views/sale_rental_views.xml",
-        "views/stock_picking_views.xml",
     ],
-    "assets": {
-        "web.assets_backend": [
-            "sale_rental_extension/static/src/js/signature_widget.js",
-            "sale_rental_extension/static/src/scss/sale_rental_kanban.scss",
-            "sale_rental_extension/static/src/xml/signature_dialog.xml",
-        ],
-    },
 }
