+++ sale_rental_extension/models/__init__.py
@@ -1,4 +1,1 @@
-from . import res_company
-from . import res_config_settings
-from . import sale_order
-from . import stock_picking
+from . import sale_order_line
