--- sale_rental_extension/models/res_company.py
@@ -1,14 +1,0 @@
-# Copyright 2026 NuoBiT Solutions SL - Eric Antones <eantones@nuobit.com>
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
-
-from odoo import fields, models
-
-
-class ResCompany(models.Model):
-    _inherit = "res.company"
-
-    rental_require_signature_validation = fields.Boolean()
-
-    rental_signature_terms = fields.Html(
-        translate=True,
-    )
