+++ sale_rental_extension/static/description/index.html
@@ -379,9 +379,8 @@ ul.auto-toc {
 <ul class="simple">
 <li>Kanban and calendar views for the Rental Positions screen</li>
 <li>Searchpanel with state filter and “Today” filter</li>
-<li>Configurable signature terms for rental delivery orders
-(Settings → Rental → Signature)</li>
-<li>Custom signature dialog with “Confirm &amp; Sign” button on rental pickings</li>
+<li>Rental quantity defaults to 1 when adding a rental product</li>
+<li>Rental quantity can be edited on confirmed orders (before pickup)</li>
 <li>Catalan and Spanish translations</li>
 </ul>
 <div class="admonition important">
