--- sale_rental_extension/static/src/js/.eslintrc.yml
@@ -1,2 +1,0 @@
-parserOptions:
-  sourceType: module
