--- sale_rental_extension/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_rental_status
