--- sale_rental_extension/views/res_config_settings_views.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2026 NuoBiT Solutions SL - Eric Antones <eantones@nuobit.com>
-     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
-<odoo>
-    <record id="res_config_settings_view_form" model="ir.ui.view">
-        <field
-            name="name"
-        >res.config.settings.view.form.inherit.sale_rental_extension</field>
-        <field name="model">res.config.settings</field>
-        <field name="inherit_id" ref="rental_base.res_config_settings_view_form" />
-        <field name="arch" type="xml">
-            <div class="app_settings_block" data-key="rental_base" position="inside">
-                <h2>Signature</h2>
-                <div class="row mt16 o_settings_container">
-                    <div class="col-12 o_setting_box">
-                        <div class="o_setting_left_pane">
-                            <field name="rental_require_signature_validation" />
-                        </div>
-                        <div class="o_setting_right_pane">
-                            <label
-                                for="rental_require_signature_validation"
-                                string="Require Signature to Validate Delivery"
-                            />
-                            <span
-                                class="fa fa-lg fa-building-o"
-                                title="Values set here are company-specific."
-                                groups="base.group_multi_company"
-                                role="img"
-                            />
-                            <div
-                                class="text-muted"
-                            >When enabled, rental delivery orders can only be validated by signing. The Validate button is replaced by the Sign button.</div>
-                        </div>
-                    </div>
-                    <div class="col-12 o_setting_box">
-                        <div class="o_setting_right_pane">
-                            <label for="rental_signature_terms" />
-                            <span
-                                class="fa fa-lg fa-building-o"
-                                title="Values set here are company-specific."
-                                groups="base.group_multi_company"
-                                role="img"
-                            />
-                            <div
-                                class="text-muted"
-                            >Terms shown in the signature dialog when signing a rental delivery order. Leave empty to use the default text.</div>
-                            <div class="content-group mt16">
-                                <field name="rental_signature_terms" widget="html" />
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </field>
-    </record>
-</odoo>
